# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"ITodo": {
			"properties": {
				"id": {
					"type": "string"
				},
				"description": {
					"type": "string"
				}
			},
			"required": [
				"id",
				"description"
			],
			"type": "object",
			"additionalProperties": true
		}
	},
	"info": {
		"title": "node-typescript",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/todo": {
			"get": {
				"operationId": "Getall",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/ITodo"
							},
							"type": "array"
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"description": {
									"type": "string"
								}
							},
							"title": "CreateBody",
							"type": "object",
							"required": [
								"description"
							]
						}
					}
				]
			}
		},
		"/todo/{id}": {
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"description": {
									"type": "string"
								}
							},
							"title": "UpdateBody",
							"type": "object",
							"required": [
								"description"
							]
						}
					}
				]
			},
			"delete": {
				"operationId": "Remove",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}